+++ packages/components/src/index.ts
@@ -12,3 +12,6 @@ export type { ICustomSelectProps } from './customSelect';
 
 export { default as CustomSuspense } from './customSuspense';
 export type { ICustomSuspenseProps } from './customSuspense';
+
+export { default as TableHeaderBtns } from './tableHeaderBtns';
+export type { ITableHeaderBtnsProps } from './tableHeaderBtns';