+++ packages/components/src/tableHeaderBtns/index.ts
@@ -1,0 +1,4 @@
+import TableHeaderBtns from './TableHeaderBtns';
+export type { ITableHeaderBtnsProps } from './TableHeaderBtns';
+
+export default TableHeaderBtns;