+++ apps/web-vite/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web-vite
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @hbwow/components@2.6.1
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @hbwow/components@2.6.0
+
 ## 0.0.18
 
 ### Patch Changes